move AssociateAccountRequest type into its own module
